tf2.pl
This is the "trak filter" program, it takes files in the /LBR directory from the lab radar, and takes all of the individual shot files, and compiles them into a single file. This is useful for tracking SNR, Velocity, or Distance. The primary index used is time.  Please note: This file is written in PERL, but it does have some shell calls, specifically one to "find" where it figures out where all of the files in the directory are. This was very convenient for me to do, as I'm doing all of my analysis with R-studio, but it may be less useful for you. This is extremely alpha, so use at your own discretion. 

So you get a file that looks like this:

Time(s) Shot 1  Shot 2  Shot 3  Shot 4  Shot 5  Shot 6  Shot 7  Shot 8  Shot 9  Shot 10 Shot 11 Shot 12 Shot 13 Shot 14 Shot 15 Shot 16 Shot 17 Shot 18 Shot 19 Shot 20
0.000000        3160.26 3178.84 3145.40 3154.90 3175.52 3146.95 3123.69 3141.82 3177.67

This is essentially just tab-separated values, so you can import all of the individual shot data, and make comparisons between shots in your analysis software.